fix random bugs, add some more (non-sensical) tests

+++ include/pyromath.h
@@ -17,7 +17,7 @@ namespace Pyro {
         float radians(float degree);
 
         static std::default_random_engine rng = std::default_random_engine(0);
-        static std::uniform_real_distribution<float> rng_dist = std::uniform_real_distribution(0.0f, 1.0f);
+        static std::uniform_real_distribution<double> rng_dist = std::uniform_real_distribution(0.0, 1.0);
 
         unsigned int random(unsigned int range);
         unsigned int random(unsigned int low, unsigned int high);
@@ -25,9 +25,9 @@ namespace Pyro {
         int random(int range);
         int random(int low, int high);
 
-        float random(); // Returns a random number between 0 and 1
-        float random(float range);
-        float random(float low, float high);
+        double random(); // Returns a random number between 0 and 1
+        double random(double range);
+        double random(double low, double high);
 
         void randomseed(unsigned int seed);
